cambios 04082022 - frank

+++ src/main/java/org/ositran/daos/InformeObraDAO.java
@@ -4,7 +4,8 @@ import com.btg.ositran.siged.domain.InformeObra;
 
 public interface InformeObraDAO {
 	public void create(InformeObra informeObra);
-	public InformeObra read(Long idInformeObra);
+	public InformeObra read(Integer idInformeObra);
 	public void update(InformeObra informeObra);
-	public void delete(Long idInformeObra);
+	public void delete(Integer idInformeObra);
+
 }
